+++ python/aws_lambda_powertools/metrics/__init__.py
@@ -1,0 +1,17 @@
+"""CloudWatch Embedded Metric Format utility
+"""
+from aws_lambda_powertools.helper.models import MetricUnit
+
+from .exceptions import MetricUnitError, MetricValueError, SchemaValidationError, UniqueNamespaceError
+from .metric import single_metric
+from .metrics import Metrics
+
+__all__ = [
+    "Metrics",
+    "single_metric",
+    "MetricUnit",
+    "MetricUnitError",
+    "SchemaValidationError",
+    "MetricValueError",
+    "UniqueNamespaceError",
+]