+++ python/aws_lambda_powertools/metrics/exceptions.py
@@ -1,0 +1,14 @@
+class MetricUnitError(Exception):
+    pass
+
+
+class SchemaValidationError(Exception):
+    pass
+
+
+class MetricValueError(Exception):
+    pass
+
+
+class UniqueNamespaceError(Exception):
+    pass